cranelift-isle: trie construction and IR cleanups
One big change here is to stop using `Term::extractor_sig`, which was
the only call that used a `TypeEnv`. However that function only uses
type information to construct the fully-qualified name of the extractor,
which is not used when building the IR. So removing it and removing the
now-unused `typeenv` parameters removes all uses of `TypeEnv` from the
`ir` and `trie` modules.

In addition, this completes the changes started in "More consistent use
of `add_inst`" (e63771f2d921cdce440620dbf38bd3e186747b1e), by always
using `add_inst` to get an `InstId`.

I also removed a number of unnecessary intermediate allocations.