Fix a flaky wait/notify test
This commit removes an assertion about the amount of time to sleep as
part of the `memory.atomic.wait*` instructions. We've already seen one
flaky test failure due to this and I think in general the timing around
it isn't super precise so it seems best to leave this specific aspect
perhaps untested for now but also pretty unlikely to regress. Otherwise
though testing specific sleeping timeout behavior isn't something we're
well structured to test at this time.

Closes #5312