x64: reduce work in non-stack using leaf functions
Cranelift has had the ability for some time to identify leaf functions;
by Cranelift's definition, a leaf function is one that knows of no other
call signatures. #1148 noted how it would be a good idea to avoid extra
frame setup work in leaf functions and #2960 implemented this for
aarch64 and s390x. This improvement was not made for x64 due to some
test failures. This change avoids any frame setup for non-stack-using
leaf functions in x64.